Обнова

+++ database/migrations/updates/2020_04_22_105308_update_210420_tables.php
@@ -48,7 +48,7 @@ class Update210420Tables extends Migration
         });
 
         Schema::table('estimates', function (Blueprint $table) {
-            $table->integer('cost')->default(0)->comment('Себестоимость')->after('number');
+            $table->decimal('cost', 10, 2)->default(0)->comment('Себестоимость')->after('number');
             $table->decimal('margin_percent', 10, 2)->default(0)->comment('Процент маржи')->after('total');
             $table->decimal('margin_currency', 10, 2)->default(0)->comment('Сумма маржи')->after('margin_percent');
